Use flat_set where it makes sense
Replaces most instances of `std::set`
and also custom sorting of `std::vector`.

2022 day12 relies on the bizarre comparison of `sorted_flat_set`,
which should die, but more work is needed.

2022 day13 fails if replacing `std::set`,
so more work needs to be done.

For 2022 day15 it makes it slightly slower.
Would probably benefit more from a `flat_map`.

2023 day17 fails to complete. More work needed
(or maybe it's just not the right container for the job).