2024 day 11 part 2
Required a whole new approach -
each stone needs to be evaluated individually,
and no stones actually need to be stored.
Instead we're just interested in the number of stones
produced by a single stone based on the remaining
number of blinks.

That solves the memory problem,
but we still need to solve the long runtime problem,
where we can use dynamic programming
in the form of a cache
since most of the results are recalculated over and over.