Specify when Dijkstra reached end using function
This is a lot more generic approach.
In many cases it reduces the amount of end nodes required to check.

However, it also made it more difficult to write overloads for
`shortest_distances_dijkstra`.
For some reason a vector of end nodes keeps deducing the `Node` type
to be `std::span<Node>`, which I don't fully understand right now.

The current best solution is to manually wrap
multiple start or end nodes into a `span`.
Hopefully this can be improved at a later point.