2022 day 22 rewrite
The old code relied on the super complicated `grid_neighbors`.
The new approach is to instead store indexes to jump locations
on edges around the actual tiles.
We need to encode the indexes based on the direction.

This should make the code a lot simpler,
and should also be easier to implement part 2 with the cube.