Refactored aoc module
Change the design from wrapping all headers into an export and hoping for the best,
to exporting only the `aoc` namespace.
We need to be careful not to include STL headers inside the module purview,
instead only include them in the global module fragment.
This fixes some warnings with <cassert> being included in module purview.
We can also include `graph.h` in the module now.

Unfortunately, this doesn't fix MSVC or GCC support, it still only works with Clang.