Better support for opting out of modules per file
Just define `AOC_DISABLE_MODULES` before includes.
This still generates the module file and creates a dependency on it
if modules were enabled for the project,
but the file can still opt out of actually using the module.

Also try to disable this for Intellisense,
which seems to struggle (at least in vscode).

Testing this revealed some minor issues.